% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{check_alert}
\alias{check_alert}
\title{Check menu for alerts}
\usage{
check_alert(menu, alerts = alerts)
}
\arguments{
\item{menu}{character string}

\item{alerts}{a vector of alerts}
}
\value{
a vector of all the relevant alerts
}
\description{
Check menu for alerts
}
