% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatsforlunch.R
\name{get_my_lunch_alerts}
\alias{get_my_lunch_alerts}
\title{See user-defined lunch alerts}
\usage{
get_my_lunch_alerts(escape = T)
}
\arguments{
\item{escape}{unescape unicode character sequences?}
}
\value{
character vector
}
\description{
See user-defined lunch alerts
}
