% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatsforlunch.R
\name{make_alert_message}
\alias{make_alert_message}
\title{Makes a message about alerts taking into account whe}
\usage{
make_alert_message(todays_alerts, upcoming)
}
\value{
character.
}
\description{
Makes a message about alerts taking into account whe
}
