% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatsforlunch.R
\name{make_greeting}
\alias{make_greeting}
\title{Makes a greeting including an adjective and the day of the week}
\usage{
make_greeting()
}
\value{
character.
}
\description{
Makes a greeting including an adjective and the day of the week
}
