% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatsforlunch.R
\name{make_time_message}
\alias{make_time_message}
\title{Makes a message commenting on the time or day}
\usage{
make_time_message()
}
\value{
character.
}
\description{
Makes a message commenting on the time or day
}
