% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatsforlunch.R
\name{set_my_lunch_alerts}
\alias{set_my_lunch_alerts}
\title{Customize lunch alerts}
\usage{
set_my_lunch_alerts(alerts, append = T)
}
\arguments{
\item{alerts}{character vector of food alerts.}

\item{append}{add to previously set alerts?}
}
\description{
Creates a character string, copies it to the clipboard and opens the .Renviron file.
Then the user pastes the string and saves the file.
}
