% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{todays_menu}
\alias{todays_menu}
\title{extract todays menu}
\usage{
todays_menu(menu, day)
}
\arguments{
\item{menu}{character string}

\item{day}{numeric weekday}
}
\value{
character string
}
\description{
extract todays menu
}
